OMAP4 CPU Thermal governor:removes repetitive info
OMAP4 CPU Thermal governor:removes repetitive info

The thermal governor prints info after each crossing of a thermal treshold.
Often treshold is crossed but thermal zone was not changed. The printed info
to clogs a console and dosen't give an effective information. The patch
removes repetitive info for all zone except for the panic zone. The
panic zone is critical and a duplicate info must be.

Change-Id: I278a938b615296264446fe7d5b209021ec0121c0
Signed-off-by: Eugen Mandrenko <ievgen.mandrenko@ti.com>